+++ test/runtests.jl
@@ -13,7 +13,7 @@ const testSCDMFunctions     = true
 const testLPBasedFunctions  = true
 const testGreyMCDMFunctions = true
 const testSummary           = true
-
+const testCopeland          = true
 
 testGreyNumbers && include("./testgreynumber.jl")
 
@@ -21,17 +21,13 @@ testUtilityFunctions && include("./testutility.jl")
 
 testSCDMFunctions && include("./testscdm.jl")
 
+testMCDMFunctions && include("./testmcdm.jl")
 
-testMCDMFunctions && let 
-    include("./testmcdm.jl")
-    include("./testcopeland.jl")
-end  
+testCopeland && include("./testcopeland.jl")
 
 testGreyMCDMFunctions && include("./testgreymcdm.jl")
 
-
 testLPBasedFunctions && include("./testlp.jl")
-
 
 testSummary && include("./testsummary.jl")